+++ vms/platformvm/txs/add_delegator_test.go
@@ -86,8 +86,8 @@ func TestAddDelegatorTxSyntacticVerify(t *testing.T) {
 		}},
 		Validator: Validator{
 			NodeID: ctx.NodeID,
-			Start:  uint64(clk.Time().Unix()),
-			End:    uint64(clk.Time().Add(time.Hour).Unix()),
+			Start:  clk.Time().Unix(),
+			End:    clk.Time().Add(time.Hour).Unix(),
 			Wght:   validatorWeight,
 		},
 		StakeOuts: stakes,
@@ -186,8 +186,8 @@ func TestAddDelegatorTxSyntacticVerifyNotAVAX(t *testing.T) {
 		}},
 		Validator: Validator{
 			NodeID: ctx.NodeID,
-			Start:  uint64(clk.Time().Unix()),
-			End:    uint64(clk.Time().Add(time.Hour).Unix()),
+			Start:  clk.Time().Unix(),
+			End:    clk.Time().Add(time.Hour).Unix(),
 			Wght:   validatorWeight,
 		},
 		StakeOuts: stakes,
