+++ vms/platformvm/txs/add_validator_test.go
@@ -85,8 +85,8 @@ func TestAddValidatorTxSyntacticVerify(t *testing.T) {
 		}},
 		Validator: Validator{
 			NodeID: ctx.NodeID,
-			Start:  uint64(clk.Time().Unix()),
-			End:    uint64(clk.Time().Add(time.Hour).Unix()),
+			Start:  clk.Time().Unix(),
+			End:    clk.Time().Add(time.Hour).Unix(),
 			Wght:   validatorWeight,
 		},
 		StakeOuts: stakes,
@@ -202,8 +202,8 @@ func TestAddValidatorTxSyntacticVerifyNotAVAX(t *testing.T) {
 		}},
 		Validator: Validator{
 			NodeID: ctx.NodeID,
-			Start:  uint64(clk.Time().Unix()),
-			End:    uint64(clk.Time().Add(time.Hour).Unix()),
+			Start:  clk.Time().Unix(),
+			End:    clk.Time().Add(time.Hour).Unix(),
 			Wght:   validatorWeight,
 		},
 		StakeOuts: stakes,
